Relay: Improved Error Channel Usage
Use `stopRead` and `stopWrite` channels to communicate
with sender / receive loops about . If channel close is initiated,
as indicated by a message to `stopRead` or `stopWrite`, don't send
errors on the error channel, which won't be read at that point anyway. Furthermore, add a small amount of buffering to the channels just to avoid possibility that a single message could inadvertently cause blocking.