WIP: partial impl for the `listen` command
This is only 30-50% representative of the code in the Go cli for this.
Still need to model the channel/greenthread stuff, actually make
requests, etc. Most of what's here is just prep work, wrangling config
values and so on.

---

Getting much closer now. Am able to use the cli to roundtrip messages
through the whole flow. Send a webhook to play, play forwards to the
client via websocket, client reads then forwards, client then takes the
response from the local server and ships it back to play.

Missing pieces:
- [x] reconnecting on connection failure.
- [x] double check the ping/pongs are happening correctly.
- [x] mimic the Go cli's error handling/shutdown signaling.
- lots of cleanup (things are messy going from a GC'd lang to a
  borrowchk'd one).
- config updates for relay (token? other stuff)